Code refactor: split the behemoth CuDensityMatOpConverter
